requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
PyExecJS==1.5.1
requests==2.28.2
six==1.16.0
urllib3==1.26.15
